+++ src/main/groovy/com/bmuschko/gradle/izpack/IzPackPlugin.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.izpack
+package com.bmuschko.gradle.izpack
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project